Add flags for red and green led threads
There is no way for the brain thread to signal to the red and green LED
threads whether the robot is to move or stay stationary. This makes it
so that the LEDs are unable to properly transition between states when
moving or stationary. Let's add some event flags to tell these threads
how to set the LEDs, so that the LEDs blink as desired depending on the
state of the robot.